+++ registry/lib/components/utils/keymap/index.ts
@@ -9,11 +9,18 @@ import { addComponentListener } from '@/core/settings'
 import { actions } from './actions'
 import { KeyBinding, KeyBindingConfig, loadKeyBindings } from './bindings'
 import { presetBase, presets } from './presets'
+import { getNumberValidator } from '@/core/utils'
 
 const options = defineOptionsMetadata({
   longJumpSeconds: {
     defaultValue: 85,
     displayName: '长跳跃秒数',
+    validator: getNumberValidator(1),
+  },
+  volumeStep: {
+    defaultValue: 10,
+    displayName: '音量调整幅度',
+    validator: getNumberValidator(1, 100),
   },
   customKeyBindings: {
     defaultValue: {} as Record<string, string>,